+++ examples/font.rs
@@ -1,10 +1,9 @@
 // Draw some sample text to the screen
-extern crate futures;
 extern crate quicksilver;
 
-use futures::future::result;
 use quicksilver::{
-    run, Asset, Future, Result, State, 
+    run, Asset, Future, Result, State,
+    combinators::result,
     geom::{Vector, Transform},
     graphics::{Color, Font, FontStyle, Image, Window, WindowBuilder}
 };